Add files via upload
What if you were told that you can leave an imprint in someone's life and be a part of their gratitude for the rest of their life! Only blood donation gives you the opportunity to add value to your life and someone else's in the most profound way! #WorldBloodDonorDay #paperboy